Fixed users creat or edit event with already exists title

+++ events/views.py
@@ -1,14 +1,26 @@
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import render, redirect, get_object_or_404
+from django.shortcuts import render, redirect
+from django.db import IntegrityError
+from django.utils.text import slugify
 
 from report.forms import ReportForm
 from report.models import Report
 from .forms import EventForm
 from .models import Event, Attendee
 
-
 # Create your views here.
+
+CATEGORIES = ['eating', 'sport', 'party']
+
+
+def check_event(request, event_category, event_slug):
+    try:
+        event = Event.objects.get(category=event_category, slug=event_slug)
+    except Event.DoesNotExist:
+        messages.warning(request, 'Event not found.')
+        return None
+    return event
 
 
 def events(request):
@@ -32,6 +44,9 @@ def events_by_category(request, event_category):
     HttpResponseObject -- event by category page
 
     """
+    if event_category not in CATEGORIES:
+        messages.warning(request, 'Invalid category.')
+        return redirect('events:feed')
     events_in_category = Event.objects.filter(category=event_category).order_by('-created_at')
     return render(request, 'events/events_by_category.html', {
         'events_in_category': events_in_category,
@@ -51,7 +66,11 @@ def new_event(request):
         event_form = EventForm(request.POST, request.FILES)
         if event_form.is_valid():
             event_title = event_form.cleaned_data['title']
-            event = event_form.save()
+            try:
+                event = event_form.save()
+            except IntegrityError:
+                messages.warning(request, f'Creation failed, {event_title} is already exists.')
+                return redirect('events:feed')
             event.user = request.user
             event.save()
             messages.success(request, f'{event_title} is created.')
@@ -68,15 +87,19 @@ def event_detail(request, event_category, event_slug):
     HttpResponseObject -- event detail page
 
     """
-    event = Event.objects.get(slug=event_slug, category=event_category)
-    if request.user.is_authenticated:
-        try:
-            joined = event.attendee_set.get(user=request.user)
-        except (KeyError, Attendee.DoesNotExist):
+    event = check_event(request, event_category=event_category, event_slug=event_slug)
+    if event:
+        event = Event.objects.get(slug=event_slug, category=event_category)
+        if request.user.is_authenticated:
+            try:
+                joined = event.attendee_set.get(user=request.user)
+            except Attendee.DoesNotExist:
+                return render(request, 'events/event_detail.html', {'event': event})
+            return render(request, 'events/event_detail.html', {'event': event, 'joined': joined})
+        else:
             return render(request, 'events/event_detail.html', {'event': event})
-        return render(request, 'events/event_detail.html', {'event': event, 'joined': joined})
     else:
-        return render(request, 'events/event_detail.html', {'event': event})
+        return redirect('events:feed')
 
 
 @login_required(login_url='/accounts/login')
@@ -87,33 +110,42 @@ def edit_event(request, event_category, event_slug):
     HttpResponseObject -- edit event page
 
     """
-    event = Event.objects.get(slug=event_slug, category=event_category)
-    if event.user == request.user and request.user.is_authenticated:
-        event_form = EventForm(initial={
-            'title': event.title,
-            'description': event.description,
-            'category': event.category,
-            'appointment_date': event.appointment_date,
-            'image_upload': event.image_upload,
+    event = check_event(request, event_category=event_category, event_slug=event_slug)
+    if event:
+        if event.user == request.user and request.user.is_authenticated:
+            event_form = EventForm(initial={
+                'title': event.title,
+                'description': event.description,
+                'category': event.category,
+                'appointment_date': event.appointment_date,
+                'image_upload': event.image_upload,
+            })
+            if request.method == 'POST':
+                event_form = EventForm(request.POST, instance=event)
+                if event_form.is_valid():
+                    if event.user == request.user and request.user.is_authenticated:
+                        event_title = event_form.cleaned_data['title']
+                        event_slug = slugify(event_title, allow_unicode=True)
+                        if Event.objects.filter(slug=event_slug).exists():
+                            messages.warning(request, f'Edition failed, {event_title} is already exists.')
+                            return redirect('events:feed')
+                        event = event_form.save()
+                        event.slug = event_slug
+                        event.save()
+                        messages.info(request, f'{event_title} was updated.')
+                    else:
+                        messages.warning(request, f'You can not edit {event.title}.')
+                    return redirect('events:feed')
+        else:
+            messages.warning(request, f'You can not edit {event.title}.')
+            return redirect('events:feed')
+        return render(request, 'events/edit_event.html', {
+            'event_form': event_form,
+            'event': event,
+            'is_edit_mode': True
         })
-        if request.method == 'POST':
-            event_form = EventForm(request.POST, instance=event)
-            if event_form.is_valid():
-                if event.user == request.user and request.user.is_authenticated:
-                    event_title = event_form.cleaned_data['title']
-                    event_form.save()
-                    messages.info(request, f'{event_title} was updated.')
-                else:
-                    messages.warning(request, f'You can not edit {event.title}.')
-                return redirect('events:feed')
     else:
-        messages.warning(request, f'You can not edit {event.title}.')
         return redirect('events:feed')
-    return render(request, 'events/edit_event.html', {
-        'event_form': event_form,
-        'event': event,
-        'is_edit_mode': True
-    })
 
 
 @login_required(login_url='/accounts/login')
@@ -124,13 +156,16 @@ def delete_event(request, event_category, event_slug):
     redirect to event feed page
 
     """
-    event = Event.objects.get(slug=event_slug, category=event_category)
-    if event.user == request.user and request.user.is_authenticated:
-        event.delete()
-        messages.warning(request, f'{event.title} is deleted.')
+    event = check_event(request, event_category=event_category, event_slug=event_slug)
+    if event:
+        if event.user == request.user and request.user.is_authenticated:
+            event.delete()
+            messages.warning(request, f'{event.title} is deleted.')
+        else:
+            messages.warning(request, f'You can not delete {event.title}.')
+        return redirect('events:feed')
     else:
-        messages.warning(request, f'You can not delete {event.title}.')
-    return redirect('events:feed')
+        return redirect('events:feed')
 
 
 @login_required(login_url='/accounts/login')
@@ -141,21 +176,24 @@ def report_event(request, event_category, event_slug):
     HttpResponseObject -- report event page
 
     """
-    event = Event.objects.get(slug=event_slug, category=event_category)
-    if request.method == 'POST':
-        report_form = ReportForm(request.POST)
-        if report_form.is_valid():
-            report = Report(
-                event=event,
-                report_type=request.POST['report_type'],
-                detail=request.POST['detail']
-            )
-            report.save()
-            messages.warning(request, f'{event.title} is reported.')
-            return redirect('events:feed')
+    event = check_event(request, event_category=event_category, event_slug=event_slug)
+    if event:
+        if request.method == 'POST':
+            report_form = ReportForm(request.POST)
+            if report_form.is_valid():
+                report = Report(
+                    event=event,
+                    report_type=request.POST['report_type'],
+                    detail=request.POST['detail']
+                )
+                report.save()
+                messages.warning(request, f'{event.title} is reported.')
+                return redirect('events:feed')
+        else:
+            report_form = ReportForm()
+        return render(request, 'events/report_event.html', {'report_form': report_form, 'event': event})
     else:
-        report_form = ReportForm()
-    return render(request, 'events/report_event.html', {'report_form': report_form, 'event': event})
+        return redirect('events:feed')
 
 
 @login_required(login_url='/accounts/login')
@@ -165,10 +203,13 @@ def joining_event(request, event_category, event_slug):
     Returns:
     HttpResponseObject -- event detail page that has join
     """
-    event = get_object_or_404(Event, category=event_category, slug=event_slug)
-    if event.attendee_set.filter(user=request.user).exists():
-        messages.warning(request, f'You have already joined {event.title}.')
+    event = check_event(request, event_category=event_category, event_slug=event_slug)
+    if event:
+        if event.attendee_set.filter(user=request.user).exists():
+            messages.warning(request, f'You have already joined {event.title}.')
+            return redirect('events:feed')
+        event.attendee_set.create(user=request.user)
+        messages.success(request, f'You have join {event.title}.')
         return redirect('events:feed')
-    event.attendee_set.create(user=request.user)
-    messages.success(request, f'You have join {event.title}.')
-    return redirect('events:feed')
+    else:
+        return redirect('events:feed')
